chore(analysis): improve static analysis compliance and documentation
- Fix 56+ PHPStan errors including null-safety, type casting, and argument mismatches.

- Update GEMINI.md and DOCUMENTATION.md with mandatory static analysis and null-safety standards.

- Refactor core models, services, and Filament resources for improved type safety.

+++ app/Filament/Widgets/AdminCommissionWidget.php
@@ -17,8 +17,8 @@ class AdminCommissionWidget extends BaseWidget
     #[Override]
     protected function getStats(): array
     {
-        $totalCommission = OrderItem::sum('commission_amount');
-        $vendorNetTotal = OrderItem::sum('vendor_net_amount');
+        $totalCommission = (float) OrderItem::sum('commission_amount');
+        $vendorNetTotal = (float) OrderItem::sum('vendor_net_amount');
 
         return [
             Stat::make(__('admin.widgets.commission.platform_commission'), '$'.number_format($totalCommission, 2))
